Changed hierarchy and moved everything to multiple files for us to work on easier.
